+++ gulpfile.js
@@ -1,35 +1,21 @@
-var pkg = require("./package.json");
-var gulp = require("gulp");
-var util = require("gulp-util");
-var del = require("del");
-var deleteLines = require("gulp-delete-lines");
-var htmlMinifier = require("html-minifier");
-var inlineTemplate = require("gulp-inline-ng2-template");
-var preprocess = require("gulp-preprocess");
-var replace = require("gulp-replace");
-var tslint = require("gulp-tslint");
-var typedoc = require("gulp-typedoc");
-var webpack = require("webpack");
-var Builder = require("systemjs-builder");
+var pkg = require("./package.json"),
+    gulp = require("gulp"),
+    deleteLines = require("gulp-delete-lines"),
+    preprocess = require("gulp-preprocess");
+
+var TASK_LINT_TYPESCRIPT = require("./build/tasks/lint-ts"),
+    TASK_CLEAN = require("./build/tasks/clean"),
+    TASK_INLINE_NG2_TEMPLATES = require("./build/tasks/inline-ng2-templates"),
+    TASK_BUNDLE_WEBPACK = require("./build/tasks/bundle-webpack"),
+    TASK_BUNDLE_SYSTEM_JS = require("./build/tasks/bundle-systemjs"),
+    TASK_DOC_API = require("./build/tasks/doc-api"),
+    TASK_INCREMENT_VERSION = require("./build/tasks/increment-version");
 
 var DIST_PATH = "./@ng2-dynamic-forms/";
 
-gulp.task("lint:modules", function () {
+gulp.task("lint:modules", TASK_LINT_TYPESCRIPT(["./modules/**/*.ts"], "./tslint.json"));
 
-    return gulp.src(["./modules/**/*.ts"], {base: "modules"})
-               .pipe(tslint({configuration: "./tslint.json"}))
-               .pipe(tslint.report());
-});
-
-
-gulp.task("clean:modules", function () {
-
-    return del([
-        DIST_PATH + "**/*",
-        "./node_modules/@ng2-dynamic-forms/**/*"
-    ]);
-});
-
+gulp.task("clean:modules", TASK_CLEAN([DIST_PATH + "**/*", "./node_modules/@ng2-dynamic-forms/**/*"]));
 
 gulp.task("prepare:modules", ["lint:modules", "clean:modules"], function () {
 
@@ -48,139 +34,42 @@ gulp.task("prepare:modules", ["lint:modules", "clean:modules"], function () {
                .pipe(gulp.dest(DIST_PATH));
 });
 
+gulp.task("inline:ng2-templates", ["prepare:modules"], TASK_INLINE_NG2_TEMPLATES([DIST_PATH + "**/*.js"], DIST_PATH));
 
-gulp.task("inline:ng2-templates", ["prepare:modules"], function () {
+gulp.task("bundle:modules:webpack:dev", ["inline:ng2-templates"], TASK_BUNDLE_WEBPACK(require("./webpack.config")));
 
-    function minify(path, ext, file, callback) {
-        try {
-            var minifiedFile = htmlMinifier.minify(file, {
-                collapseWhitespace: true,
-                caseSensitive: true,
-                removeComments: true,
-                removeRedundantAttributes: true
-            });
+gulp.task("bundle:modules:webpack:prod", ["inline:ng2-templates"], TASK_BUNDLE_WEBPACK(require("./webpack.min.config")));
 
-            callback(null, minifiedFile);
-        }
-        catch (err) {
-            callback(err);
-        }
-    }
+gulp.task("bundle:modules:systemjs", ["inline:ng2-templates"], TASK_BUNDLE_SYSTEM_JS());
 
-    return gulp.src([DIST_PATH + "**/*.js"], {base: DIST_PATH})
-               .pipe(deleteLines({
-                   'filters': [/moduleId: module.id/]
-               }))
-               .pipe(inlineTemplate({
-                   base: DIST_PATH,
-                   removeLineBreaks: true,
-                   target: "es5",
-                   templateProcessor: minify,
-                   useRelativePaths: true
-               }))
-               .pipe(gulp.dest(DIST_PATH));
-});
+gulp.task("prime:modules", ["bundle:modules:webpack:dev", "bundle:modules:webpack:prod"], function () {
 
-
-gulp.task("bundle:modules:webpack", ["inline:ng2-templates"], function (callback) {
-
-    webpack(require("./webpack.config"), function (error, stats) {
-
-        if (error) {
-            throw new util.PluginError("webpack", error);
-        }
-
-        util.log("bundle:modules", stats.toString({
-            chunks: false,
-            colors: true,
-            hash: false,
-            version: false
-        }));
-        callback();
-    });
-});
-
-
-gulp.task("bundle:modules:systemjs", ["inline:ng2-templates"], function (callback) {
-
-    var builder = new Builder({
-
-        map: {
-            "modules": "modules"
-        },
-        packages: {
-            "modules": {
-                defaultExtension: "js"
-            }
-        }
-    });
-
-    builder.buildStatic("modules/core/index.js", "bundles/core.umd.js", {
-
-        externals: [
-            "@angular/common",
-            "@angular/compiler",
-            "@angular/core",
-            "@angular/forms",
-            "@angular/http",
-            "@angular/platform-browser",
-            "@angular/platform-browser-dynamic",
-            "@angular/router",
-            "@angular2-material/checkbox",
-            "@angular2-material/core",
-            "@angular2-material/input",
-            "@angular2-material/radio",
-            "@ng2-dynamic-forms/core"
-        ]
-    }).then(function () {callback();});
-});
-
-
-gulp.task("prime:modules", ["lint:modules", "clean:modules", "prepare:modules", "inline:ng2-templates", "bundle:modules:webpack"], function () {
-
-    return gulp.src([DIST_PATH + "**/*.umd.js",], {base: "@ng2-dynamic-forms"})
+    return gulp.src([DIST_PATH + "**/*.umd.js", DIST_PATH + "**/*.umd.min.js",], {base: "@ng2-dynamic-forms"})
                .pipe(deleteLines({'filters': [/# sourceMappingURL=/]}))
                .pipe(gulp.dest("./node_modules/@ng2-dynamic-forms/"))
                .pipe(gulp.dest(DIST_PATH));
 });
 
+gulp.task("build:documentation", TASK_DOC_API(["./modules/*/src/**/*.ts"],
+    {
+        exclude: "./modules/**/*.spec.ts",
+        module: "commonjs",
+        target: "es5",
+        out: "./docs/",
+        name: "ng2 Dynamic Forms",
+        includeDeclarations: true,
+        ignoreCompilerErrors: true
+    }
+));
 
-gulp.task("build:documentation", function () {
-
-    return gulp.src(["./modules/*/src/**/*.ts"], {read: false})
-               .pipe(typedoc({
-                   exclude: "./modules/**/*.spec.ts",
-                   module: "commonjs",
-                   target: "es5",
-                   out: "./docs/",
-                   name: "ng2 Dynamic Forms",
-                   includeDeclarations: true,
-                   ignoreCompilerErrors: true
-               }));
-});
-
-
-gulp.task("increment:version", function () {
-
-    var versionNumber = Number(pkg.version.slice(-2)),
-        versionString = /(\d\.\d\.\d-[a-z]+\.)\d{1,2}/,
-        newVersionNumber = versionNumber + 1,
-        newVersionString = pkg.version.replace(versionString, "$1" + newVersionNumber),
-        versionField = /("version":\s)"\d\.\d\.\d-[a-z]+\.\d{1,2}"/,
-        dependencyField = /("@ng2-dynamic-forms\/[a-z\-]+":\s)"\^\d\.\d\.\d-[a-z]+\.\d{1,2}"/g;
-
-    return gulp.src(["./package.json", "./modules/**/package.json"], {base: "./modules"})
-               .pipe(replace(versionField, "$1" + '"' + newVersionString + '"'))
-               .pipe(replace(dependencyField, "$1" + '"^' + newVersionString + '"'))
-               .pipe(gulp.dest("./modules"));
-});
-
+gulp.task("increment:version", TASK_INCREMENT_VERSION(pkg, ["./package.json", "./modules/**/package.json"], "./modules"));
 
 gulp.task("build:modules", [
     "lint:modules",
     "clean:modules",
     "prepare:modules",
     "inline:ng2-templates",
-    "bundle:modules:webpack",
+    "bundle:modules:webpack:dev",
+    "bundle:modules:webpack:prod",
     "prime:modules"
 ]);